requirements.txt
Django>=5.2,<5.3
wagtail>=7.2,<7.3
python-dotenv

Pillow